+++ packages/fighting-design/link/src/interface.d.ts
@@ -10,4 +10,4 @@ export type { LinkProps } from './props'
 export type LinkTarget = '_self' | '_blank' | '_parent' | '_top'
 
 /** 不同 hover 状态 */
-export type LinkState = 'line' | 'bag'
+export type LinkState = 'line' | 'background'