+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - `f-progress` 组件 `type` 新增 `info` 和 `default` 配置项
 - `f-swap` 组件 `type` 废除 `default` 配置项
+- `f-link` 组件 `state` 配置项的 `bag` 可选值改为 `background`
+- `f-link` 组件废除 `beforeIcon` 和 `afterIcon` 插槽
 
 ## 0.23.1 (2023-02-16)